Write the current date when creating a PDF
I have an Emacs extension that would automatically replace the date
with the current date every time I saved the tex file. However, lately
I have switched to using vscode for editing LaTeX files. Thus, the
time was not being updated.

I changed this to use the current date whenever creating the PDF file.
It was nice to know that the date was fixed to the last time I actually
edited the file, not when I built it. However, it's not a good enough
feature to bother with trying to keep the date correct. Most of the
times I need to build the PDF, I will likely update the LaTeX anyway.
And if I don't, I've at least decided it is up to date enough.